[enh #76] enum関連をTitleSceneクラス内に移動
